Rename test filenames as per PEP8
